improved list display fixed unfollow script, added flag for ready confirmation.
still need to fix and improve/implement:

wrong botStatus after script is run. -script complete successfully but shows an error message.

ignore users list - next to minus button (remove) a plus (+) button that will add the username to a list. there on the next compare this user wont be included on the unfollow list.

search bar for unfollow list